improved function to separate language pairs using line breaks
Until now, a line break was being inserted after every three words. But this would mess things up when specializations were mentioned, since they

This has been fixed by adding a list of languages with which a match is performed to determine whether a word is a language name. Conditions have
Now a line break is appears after every language pair and every set of specializations.